+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add support for `Value.TypeOf()` corresponding to JavaScript `typeof` operator.
+
 ### Changed
 
 ## [v0.33.0] - 2025-05-15